@@ -1,16 +1,15 @@
-﻿namespace DatingApp.API.Errors
-{
-    public class ApiException
-    {
-        public int StatusCode { get; set; }
-        public string Message { get; set; }
-        public string? Details { get; set; }
+﻿namespace DatingApp.API.Errors;
 
-        public ApiException(int statusCode, string message, string? details)
-        {
-            StatusCode = statusCode;
-            Message = message;
-            Details = details;
-        }
+public class ApiException
+{
+    public ApiException(int statusCode, string message, string? details)
+    {
+        StatusCode = statusCode;
+        Message = message;
+        Details = details;
     }
+
+    public int StatusCode { get; set; }
+    public string Message { get; set; }
+    public string? Details { get; set; }
 }
