@@ -2,24 +2,24 @@ import { Injectable } from '@angular/core';
 import { CanActivate } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
 import { map, Observable } from 'rxjs';
-import { AuthService } from '../_services/auth.service';
+import { AccountService } from '../_services/account.service';
 
 @Injectable({
-  providedIn: 'root',
+  providedIn: 'root'
 })
 export class AuthGuard implements CanActivate {
-  constructor(
-    private authService: AuthService,
-    private toastr: ToastrService
-  ) {}
+  constructor(private accountService: AccountService, private toastr: ToastrService) {}
 
   canActivate(): Observable<boolean> {
-    return this.authService.currentUser$.pipe(
+    return this.accountService.currentUser$.pipe(
       map(user => {
         if (user) return true;
-        this.toastr.error('You shall not pass!');
-        return false;
+        else {
+          this.toastr.error('You shall not pass!');
+          return false
+        }
       })
-    );
+    )
   }
+  
 }
