@@ -1,17 +1,16 @@
 import { Component, OnInit } from '@angular/core';
-import { UserSession } from './_models/UserSession';
-import { AuthService } from './_services/auth.service';
+import { User } from './_models/user';
+import { AccountService } from './_services/account.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+  styleUrls: ['./app.component.css']
 })
 export class AppComponent implements OnInit {
-  title = 'DatingApp-SPA';
-  users: any;
+  title = 'Dating app';
 
-  constructor(private authService: AuthService) {}
+  constructor(private accountService: AccountService) {}
 
   ngOnInit(): void {
     this.setCurrentUser();
@@ -19,10 +18,10 @@ export class AppComponent implements OnInit {
 
   setCurrentUser() {
     const userString = localStorage.getItem('user');
-
     if (!userString) return;
-
-    const user: UserSession = JSON.parse(userString);
-    this.authService.setCurrentUser(user);
+    const user: User = JSON.parse(userString);
+    this.accountService.setCurrentUser(user);
   }
+
+
 }
