@@ -1,26 +1,17 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
 
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
-  styleUrls: ['./home.component.css'],
+  styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
   registerMode = false;
   users: any;
 
-  constructor(private userService: UserService) {}
+  constructor() { }
 
   ngOnInit(): void {
-    this.userService.getUsers().subscribe({
-      next: response => {
-        this.users = response;
-      },
-      error: error => {
-        console.log(error);
-      },
-    });
   }
 
   registerToggle() {
@@ -29,5 +20,6 @@ export class HomeComponent implements OnInit {
 
   cancelRegisterMode(event: boolean) {
     this.registerMode = event;
-  }
+  } 
+
 }
