@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using DatingApp.Models.Database.DataModel;
-using DatingApp.Models.Dtos.Photo;
+using DatingApp.Models.Dtos;
 
 namespace DatingApp.BLL.Mappers;
 
@@ -8,8 +8,6 @@ public class PhotoProfile : Profile
 {
     public PhotoProfile()
     {
-        CreateMap<Photo, PhotoForDetailsDto>(MemberList.None);
-        CreateMap<PhotoForCreationDto, Photo>(MemberList.None);
-        CreateMap<Photo, PhotoForReturnDto>(MemberList.None);
+        CreateMap<Photo, PhotoDto>();
     }
 }
