@@ -1,11 +1,10 @@
-﻿using DatingApp.Models.Dtos.Photo;
+using CloudinaryDotNet.Actions;
+using Microsoft.AspNetCore.Http;
 
 namespace DatingApp.BLL.Services.Interfaces;
 
 public interface IPhotoService
 {
-    Task<PhotoForReturnDto> UploadPhotoAsync(int userId, PhotoForCreationDto photo);
-    Task<PhotoForReturnDto> GetPhotoAsync(int id);
-    Task<PhotoForReturnDto> SetPhotoAsMainAsync(int userId, int photoId);
-    Task RemoveUserPhotoAsync(int userId, int photoId);
+    Task<ImageUploadResult> AddPhotoAsync(IFormFile file);
+    Task<DeletionResult> DeletePhotoAsync(string publicId);
 }
