@@ -4,12 +4,12 @@ namespace DatingApp.DAL.Helpers;
 
 public class PagedList<T> : List<T>
 {
-    public PagedList(List<T> items, int count, int pageNumber, int pageSize)
+    public PagedList(IEnumerable<T> items, int count, int pageNumber, int pageSize)
     {
-        TotalCount = count;
-        PageSize = pageSize;
         CurrentPage = pageNumber;
         TotalPages = (int)Math.Ceiling(count / (double)pageSize);
+        PageSize = pageSize;
+        TotalCount = count;
         AddRange(items);
     }
 
@@ -18,11 +18,11 @@ public class PagedList<T> : List<T>
     public int PageSize { get; set; }
     public int TotalCount { get; set; }
 
-    public static async Task<PagedList<T>> CreateAsync(IQueryable<T> source, int pageNumber, int pageSize)
+    public static async Task<PagedList<T>> CreateAsync(IQueryable<T> source,
+        int pageNumber, int pageSize)
     {
         var count = await source.CountAsync();
         var items = await source.Skip((pageNumber - 1) * pageSize).Take(pageSize).ToListAsync();
-
         return new PagedList<T>(items, count, pageNumber, pageSize);
     }
 }
