@@ -1,14 +1,15 @@
-﻿namespace DatingApp.Models.Database.DataModel;
+using System.ComponentModel.DataAnnotations.Schema;
 
-public class Photo : IEntity
+namespace DatingApp.Models.Database.DataModel;
+
+[Table("Photos")]
+public class Photo
 {
+    public int Id { get; set; }
     public string Url { get; set; }
-    public string Description { get; set; }
-    public DateTime DateAdded { get; set; }
     public bool IsMain { get; set; }
     public string? PublicId { get; set; }
 
-    public User User { get; set; }
-    public int UserId { get; set; }
-    public int Id { get; set; }
+    public int AppUserId { get; set; }
+    public AppUser AppUser { get; set; }
 }
